requirements.txt
# --- Flask & Core ---
Flask==3.1.3
flask-cors==6.0.2
python-dotenv==1.2.2
Werkzeug==3.1.8

# --- Database ---
supabase==2.28.3

# --- Cryptography ---
cryptography==46.0.6

# --- OCR (Hybrid: Local preprocessing + Gemini Vision API) ---
pytesseract>=0.3.10        # Lightweight wrapper for Tesseract binary (~10KB)
opencv-python-headless==4.8.1.78  # Image preprocessing only, no GUI (~30MB)
numpy<2.0                  # Pinned: opencv 4.8 not compatible with numpy 2.x
Pillow>=10.0.0

# --- LLM Field Extraction ---
google-generativeai==0.8.6  # Gemini Flash Vision — structured extraction (~2MB)

# --- PDF Support ---
pypdfium2>=5.0.0           # Lightweight PDF page extraction (~5MB)
